Move getWithLoader to a new file
